Save and provide access to generator context
This commit makes it possible to view the components which were used for
generation of the fake data. It saves the results of the recursive
formatter calls alongside with the formatter names into dict and
provides access to this dict for library users through the 'parts'
attribute of the returned string.